# pnmc-api/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PNMC Migration API
  version: 1.0.0
  summary: Contrato API inicial para migracion incremental PNMC
  description: |
    Contrato base para desacoplar integraciones sensibles del frontend actual,
    habilitando una migracion paralela sin tocar la UI existente.
servers:
  - url: http://localhost:8080
    description: Local
  - url: https://api-dev.pnmc.local
    description: Dev (referencia)
...
paths:
...
  /api/v1/agenda/events:
    get:
      tags: [agenda]
      summary: Listar eventos de agenda
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - in: query
          name: month
          schema:
            type: string
            pattern: '^[0-9]{4}-(0[1-9]|1[0-2])$'
          description: Filtro por mes (YYYY-MM)
        - in: query
          name: tag
          schema:
            type: string
          description: Filtro por etiqueta
      responses:
        '200':
          description: Eventos retornados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedAgendaResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/agenda/events/{eventId}:
    get:
      tags: [agenda]
      summary: Obtener detalle de evento
      parameters:
        - $ref: '#/components/parameters/EventId'
      responses:
        '200':
          description: Evento encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgendaEvent'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/admin/data/agenda/events:
    post:
      tags: [admin-data]
      summary: Crear o actualizar evento de agenda en base backend
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgendaEventUpsertRequest'
      responses:
        '200':
          description: Evento persistido
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  parameters:
    Limit:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Offset:
      in: query
      name: offset
      schema:
        type: integer
        minimum: 0
        default: 0
    EventId:
      in: path
      name: eventId
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Solicitud invalida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    InternalError:
      description: Error interno
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
...
    ProblemDetails:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
    Pagination:
      type: object
      required: [limit, offset, total]
      properties:
        limit: { type: integer }
        offset: { type: integer }
        total: { type: integer }
    AgendaEvent:
      type: object
      required: [id, title, date]
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
        category: { type: string }
        date:
          type: string
          format: date
        timeLabel: { type: string }
        location: { type: string }
        municipality: { type: string }
        department: { type: string }
        organizer: { type: string }
        imageUrl: { type: string }
        tags:
          type: array
          items: { type: string }
        source:
          type: string
          enum: [backend_sql]
    PagedAgendaResponse:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AgendaEvent'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    AgendaEventUpsertRequest:
      type: object
      required: [title]
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
        category: { type: string }
        date: { type: string }
        timeLabel: { type: string }
        location: { type: string }
        municipality: { type: string }
        department: { type: string }
        organizer: { type: string }
        imageUrl: { type: string }
        tags:
          type: array
          items: { type: string }
...